backend/pyproject.toml
[project]
name = "project_hermes"
version = "0.1.0"
description = "Project Hermes - Multi-Agent Travel Planning System"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.177.0,<1.0.0",
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.6.1",
    "pytest>=8.3.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.1",
    "langchain-google-genai>=0.0.3",
    "langchain-anthropic>=0.1.1",
    "google-generativeai>=0.3.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
    "pytest",
    "mypy",
    "ruff",
]

[project.scripts]
kickoff = "project_hermes.main:kickoff"
run_crew = "project_hermes.main:kickoff"
plot = "project_hermes.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
